pyproject.toml
[project]
name = "template-version-check"
license = {file = "LICENSE"}
readme = "README.md"
authors = [
    { name = "LINCC Frameworks", email = "lincc-frameworks-team@lists.lsst.org" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
dynamic = ["version"]
dependencies = [
    "pyyaml",
    "GitPython",
    "packaging",
]

# On a mac, install optional dependencies with `pip install '.[dev]'` (include the single quotes)
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov", # Used to report total code coverage
    "pre-commit", # Used to run checks before finalizing a git commit
    "sphinx==6.1.3", # Used to automatically generate documentation
    "sphinx_rtd_theme==1.2.0", # Used to render documentation
    "sphinx-autoapi==2.0.1", # Used to automatically generate api documentation
    "black", # Used for static linting of files
]

[build-system]
requires = [
    "setuptools>=45", # Used to build and package the Python project
    "setuptools_scm>=6.2", # Gets release version from git. Makes it available programmatically
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/check_template_version/_version.py"

[tool.black]
line-length = 110


